feat: add version dropdown
